add expire record

+++ src/mongo/index.ts
@@ -9,31 +9,42 @@ interface Record extends RecordModel {
     _id: string
 }
 
+interface GetRecordResponse extends Pick<RecordModel, 'text'> {
+    isOneTime: boolean
+}
+
 const maxTtlSeconds = 5 * 60
 const maxSaveAttempts = 10
+const oneTimePrefix = 'o'
+const expirationPrefix = 'e'
 
-// TODO: Return here only what needed to show to reduce used traffic
-export async function getRecord(id: string): Promise<RecordModel | null> {
-    const collection = await getCollection()
+export async function getRecord(id: string): Promise<GetRecordResponse | null> {
+    const recordIsOneTime = isOneTime(id)
+    const collection = await getCollection(recordIsOneTime)
 
     // TODO: realm-web have bug and findOneAndDelete isn't working as intended.
     // https://github.com/realm/realm-js/issues/6497
     const record = await collection.findOne({
         _id: id,
+    }, {
+        projection: {
+            text: 1,
+        },
     })
 
     if (!record) {
         return null
     }
 
-    await collection.deleteOne({
-        _id: id,
-    })
+    if (recordIsOneTime) {
+        await collection.deleteOne({
+            _id: id,
+        })
+    }
 
     return {
-        id: record._id,
         text: record.text,
-        expireAt: record.expireAt,
+        isOneTime: recordIsOneTime,
     }
 }
 
@@ -41,8 +52,6 @@ export async function createRecord(record: Pick<RecordModel, 'text' > & {
     ttlSeconds: number
     isOneTime: boolean
 }) {
-    // TODO: isOneTime will be different by non-onetime by storing documents to different collection
-    // and querying by findOne and key to understand one-time / non-one-time will be character in key
     if (record.ttlSeconds > maxTtlSeconds) {
         throw new Error('invalid ttl')
     }
@@ -50,12 +59,12 @@ export async function createRecord(record: Pick<RecordModel, 'text' > & {
     const expireAt = new Date()
     expireAt.setSeconds(expireAt.getSeconds() + record.ttlSeconds)
 
-    const collection = await getCollection()
+    const collection = await getCollection(record.isOneTime)
 
     let attempts = 0
     while (true) {
         try {
-            const id = generateId()
+            const id = generateId(record.isOneTime)
 
             await collection.insertOne({
                 _id: id,
@@ -77,16 +86,46 @@ export async function createRecord(record: Pick<RecordModel, 'text' > & {
 }
 
 
-async function getCollection() {
-    const client = await getDatabase()
-
-    return client.collection<Omit<Record, 'id'>>('records')
+function getCollection(isOneTime: boolean) {
+    return isOneTime
+        ? getOneTimeCollection()
+        : getExpireCollection()
 }
 
-function generateId() {
+async function getExpireCollection() {
+    const client = await getDatabase()
+
+    return client.collection<Record>('records')
+}
+
+async function getOneTimeCollection() {
+    const client = await getDatabase()
+
+    return client.collection<Record>('one-time-records')
+}
+
+function generateId(isOneTime: boolean) {
     const { env } = getRequestContext()
 
     const array = crypto.getRandomValues(new Int8Array(4))
     const hashids = new Hashids(env.HASHIDS_SALT)
-    return hashids.encode(Array.from(array.map(n => Math.abs(n))))
+    const id = hashids.encode(Array.from(array.map(n => Math.abs(n))))
+
+    const prefix = isOneTime
+        ? oneTimePrefix
+        : expirationPrefix
+
+    return prefix + id
+}
+
+function isOneTime(id: string) {
+    if (id.startsWith(oneTimePrefix)) {
+        return true
+    }
+
+    if (id.startsWith(expirationPrefix)) {
+        return false
+    }
+
+    throw new Error(`Unknown id format. Id: ${id}`)
 }
